Move decl/ext @implementation tests to attr
Turns out that @implementation will be applied to things other than extensions, and also that it requires many individual test files.
