+++ test/attr/attr_implementation/Inputs/objc_class_extension.h
