+++ test/attr/attr_implementation/Inputs/objc_class_extension_internal.h
@@ -1,4 +1,4 @@
-#import "objc_implementation_class_extension.h"
+#import "objc_class_extension.h"
 
 @interface ObjCClass ()
