+++ test/attr/attr_implementation/cdecl.swift
