+++ test/attr/attr_implementation/cdecl_early_adopter.swift
