+++ test/attr/attr_implementation/objc_early_adopter.swift
