--- test/decl/ext/Inputs/objc_implementation_class_extension.modulemap
@@ -1,9 +1,0 @@
-module objc_implementation_class_extension {
-  header "objc_implementation_class_extension.h"
-  export *
-}
-
-module objc_implementation_class_extension_internal {
-  header "objc_implementation_class_extension_internal.h"
-  export *
-}
